requirements.txt
google-generativeai>=0.8.0
neo4j>=5.19.0
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.35.0
pyvis>=0.3.2
tenacity>=8.3.0
